requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
importlib_resources==6.4.0
kiwisolver==1.4.5
matlabengine==24.1.2
matplotlib==3.9.1
mpmath==1.3.0
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.13.1
zipp==3.19.2
